// frontend/src/app/core/models/usuario.model.ts
export interface Usuario {
  id_usuario: number;
  nombre: string;
  apellido?: string;
  email: string;
  rol: 'admin' | 'ejecutivo' | 'operador' | 'analista' | 'cliente';
  telefono?: string;
}

export interface LoginRequest { email: string; password: string; }

export interface RegisterRequest {
  nombre: string;
  apellido?: string;
  email: string;
  password: string;
  rol?: string;
  rut?: string;
  telefono?: string;
  tipo_cliente?: 'B2C' | 'B2B';
  razon_social?: string | null;
}

export interface AuthResponse { token: string; user?: Usuario; usuario?: Usuario; }